+++ crates/cli/tests/dynamic-linking-scripts/javy-json-id.js
@@ -1,1 +1,0 @@
-console.log(Javy.JSON.toStdout(Javy.JSON.fromStdin()));